Minor bug fixes, as well as a port of TraceEvent fix for manifest that are broken into pieces but and are comming from mulitple processes (thus they interleve)

+++ src/TraceEvent/TraceLog.cs
@@ -753,7 +753,7 @@ namespace Microsoft.Diagnostics.Tracing.Etlx
             if (eventRecord->ExtendedDataCount == 1)
             {
                 int idIndex = (int)eventRecord->ExtendedData;
-                if (idIndex < relatedActivityIDs.Count)
+                if ((uint) idIndex < (uint) relatedActivityIDs.Count)
                     return relatedActivityIDs[idIndex];
             }
             return Guid.Empty;
